storage: Refactor CollectionManager to spawn a task per-collection (#21116)
This PR updates the `CollectionManager` in `storage-controller` to spawn
an individual `tokio::task` per managed collection.

Previously `CollectionManager` spawned a _single_ `tokio::task` that
pulled updates off of a channel for all of the registered collections,
and would append these in a single batch. To rate limit writes to
persist for the webhook source, we added artificial latency to the write
in https://github.com/MaterializeInc/materialize/pull/20885, this had
two side effects:

1. All writes were rate limited, regardless of the collection they were
writing to. If `Collection(A)` received an update at `t`, then
`Collection(B)` received an update at `t + 100ms`, the update for
`Collection(B)` would have to wait until `t + 1,000ms` until it could
proceed.
2. Bumping uppers for all registered collections ran as a separate
`Future` that was in a `select!` with the append `Future`. This meant
that if we received an update for `Collection(A)` at `t` and no updates
for `Collection(B)`, it would be _at least_ until `t + 1000ms` until the
upper for `Collection(B)` would be bumped.

This PR refactors the `CollectionManager` so each collection we're
appending to has it's own "run loop", which solves the aforementioned
issues and IMO is a better implementation than what we have previously
because:

1. Removes complexity around batching updates for multiple collections,
since each "task" handles updates for a single collection.
2. Removes the complexity around unregistering a collection and waiting
for a "notify". Now we can shutdown that collections "run loop" and wait
for it to exit.
3. Makes it much simpler to change the interval/tick/append rate for an
individual collection, e.g. supporting `TIMESTAMP INTERVAL`.

### Motivation

* This PR fixes a previously unreported bug.
We recently discovered excess latency in some of our `BuiltinSource`s,
[Slack](https://materializeinc.slack.com/archives/CU7ELJ6E9/p1691616298456569).
I'm pretty sure this is caused by the rate limiting that was recently
added for webhooks.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - N/a