adapter: Support SHOW COLUMNS as the introspection user
SHOW COLUMNS is special. The other SHOW commands are either
showing information about a single object and don't allow arbitrary
extensions to the underlying query (like SHOW TABLE) or are general
queries that don't need to refer to any particular object (like SHOW
INDEXES). In order to allow SHOW COLUMNS to refer to user objects
and accept arbitrary where conditions we have to special case it.