adapter,compute: make `ReplicaId` an enum
This commit converts the `ReplicaId` type from a plain `u64` into an
enum that has `User` and `System` variants. This brings replica IDs in
line with cluster IDs and makes it possible to differenciate between
user and system replicas based on their IDs.

There are no breaking changes to the system tables, since we already
changed the type of `mz_cluster_replicas.id` in a0714eb. The only
user-visible change is that replicas in system clusters now have `s*`
IDs instead of `u*` IDs.

The names of replica processes in the orchestrator have changed. We will
need to adjust for this in our testing and cloud infrastructure in.
Also, the durable catalog state requires a migration. Both of these
tasks are deferred to follow-up commits.

# Conflicts:
#	src/adapter/src/catalog.rs
#	src/adapter/src/catalog/builtin_table_updates.rs
#	src/adapter/src/catalog/storage.rs
#	src/adapter/src/coord/sequencer/inner.rs
#	src/cluster-client/src/lib.rs
#	src/compute-client/src/controller.rs
#	test/sqllogictest/audit_log.slt