persist: split persistcli into a new crate
In #20954, we get a circular dep from mz_persist_client (dev-deps) to
mz_persist_txn to mz_persist_client. Apparently this is not a great
idea, even though cargo allows it. Concretely, my r-a seems to be
crashing noticeably less when I comment out the circular dep in the txn
branch.

Context at https://github.com/rust-lang/rust-analyzer/issues/14167

We've got another one of these in mz_expr and mz_expr_test_util, but
I'll leave that for someone else.