adapter: teach handle_startup to not depend on session
Pass all needed fields to handle_startup so that in the future it does
not need to depend on a passed session.

Refactor various session-using things to use a ConnMeta instead, which
has taken over the role of the SessionMeta trait. This will allow for
coord+adapter to exchange only connection ids in the future.