Vectors only own allocation if capacity larger 0
Prior to this change, we were counting empty vectors as allocations for
arrangement size logging. This is misleading as vectors without a
capacity do not have a heap allocation, and as such should not report
any allocation. This change fixes this by only calling the callback if
the vector owns an allocation. This behavior is correct as the callback
should be called for each owned allocation.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>