adapter: stop sending sessions during startup
Change startup flow to better handle errors.

Also a notable change to the audit log when auto creating users: the
user field of the version event will now be NULL, indicating the system
initiated the change. The user name is still present in the event
details though, so we don't lose any information.