mz: add retry to the disable command (#21250)
This pull request (PR) adds the `Retry` feature to the `disable region`
command, similar to the `enable region` command.
The underlying `delete_region` method in the `cloud-client` already
contains a loop for handling region disablement. However, it may fail in
the case of an unknown error, such as the 502 error.