storage: extract antichain time iff T: TotalOrder
The places touched by this commit are attempting to extract a single
timestamp from an Antichain which only makes sense when the timestamp is
totally ordered. However the code as written would compile fine even if
the `TotalOrder` bound were to be lifted, leading to a logic bug.

Using `Antichain::as_option` which is only defined for `T: TotalOrder`
ensures that we get a compile error if we ever attempt to remove the
trait bound.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>