CODEOWNERS: add DevEx to `/src/sql-parser`
Most changes that impact external tooling owned by @MaterializeInc/devex (_e.g._ dbt, Terraform) are changes to the SQL syntax. Adding our team as a code owner of `/src/sql-parser` to increase awareness of ongoing work. This isn't meant to set any precedence WRT code reviews, and might be reverted if it turns out to be more overwhelming than useful.

Aligned with @chaas to make sure the Surfaces team is aware of this change.