sqlsmith: Silence errors around the use of the avg_internal_v1() function
The avg_internal_v1() function was recently introduced to preserve legacy
behavior of aggregations. It now shows up in the list of functions and
SQLSmith considers it to be non-aggregate, using it in places within
the query where it is not allowed.

Until the issue is fixed on the SQLSmith side, we will silence the
error on the mzcompose.py side.

Relates to #21295