Make statement logging persistent
This stores the statement log in storage-managed collections, rather
than in system tables, so they won't be refreshed and cleared on
startup. In order to avoid unnecessarily high load on Cockroach, we
buffer the updates and flush them every 5 seconds.

On startup, we re-ingest the log in order to truncate the parts of it
that are older than a configurable value, by default 30
days. Currently, this has the potential to use significant memory; we
may need to switch to a more incremental approach. We will need to
watch this closely when turning on statement logging in prod.