storage: gracefully handle shard finalization in persist_sink
Our current approach to dropping a source results in a race between
shard finalization (which happens in the controller) and dataflow
shutdown (which happens in clusterd). If a source is created and dropped
fast enough -or the two commands get sufficiently delayed- then it's
possible for clusterd to attempt to start writing to a shard that the
controller has already finalized.

This resulted in a panic in clusterd instead of a graceful shutdown of
the dataflow.

This PR implements a solution to this problem by restarting the
ingestion dataflow whenever an upper mismatch is encountered in the
storage persist sink implementation. This is a stopgap solution until
concurrent ingestion is propertly implemented and such upper mismatches
get handled inside the operators.

Fixes #21314

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>