adapter: thread otel context throug internal command/response channels
Before, the OpenTelemetry context would be lost when traversing a
command/response queue. Now we try and keep that context in more places.

This also does change a few messages where before we were passing a Span
through a channel to now instead pass an otel context and re-attach that
as a parent.