sql[-parser]: support `TABLE` expressions
The `TABLE` expression is an esoteric SQL feature that retrieves all
rows from the named SQL table, as in:

    TABLE db.schm.tab

It is not widely used, but is nonetheless important to support for
PostgreSQL parity.

This commit adds support and documentation for this expression. It also
adds documentation for `VALUES` expressions, which I happened to notice
were missing while working on the `TABLE` expression documentation.