adapter: remove Command::DumpCatalog
We can use a catalog snapshot similar to other catalog-only methods.
This no longer allows rbac to verify that a superuser is doing the dump.
However, because we only expose this over the internal port which is
auth'd as mz_system (a superuser), there's no actual permissions
difference to not checking this. No other commands were using the rbac
check command fn, so we're able to remove that and a few other things in
the chain. Arguably some other commands *should* have been using this
(GetSystemVar, SetSystemVar), but were not. All of these privileged APIs
were already limited to privileged use and aren't accessible via SQL.