adapter,sql: fill in pg_type.typreceive
The `pg_type.typreceive` column specifies the name of the internal
PostgreSQL function that can decode binary-encoded values of the type.
We were previously always filling this column in as NULL, but doing so
trips up Power BI (see #11158). This commit fills in the `typreceive`
column as PostgreSQL does for all builtin types.

Fix #11158.

Co-authored-by: Matt Jibson <matt.jibson@gmail.com>