compute-client: add mz_dataflow_initial_output_duration_seconds metric
This commit adds the `mz_dataflow_initial_output_duration_seconds`
metric back in, this time on the controller side. The metric is
per-replica so maintaining it within the replica tasks is easiest. For
this the replica task needs to start observing compute commands and
responses more closely, to get to know about new collections being
created and their frontiers advancing beyond their `as_of`s.