scalability framework: don't use Cursor from multiple tasks concurrently
Cursor is _not_ thread safe but the tasks in the Thread pool would
sometimes use their cursor concurrently with other tasks, based on
timing.

This would sometimes lead to exceptions when trying to access `ntuples`
within `fetchall().

For reference, this is `Cursor.fetchall()`:

```
def fetchall(self) -> List[Row]:
    """
    Return all the remaining records from the current recordset.

    :rtype: Sequence[Row], with Row defined by `row_factory`
    """
    self._fetch_pipeline()
    self._check_result_for_fetch()
    assert self.pgresult
    records = self._tx.load_rows(self._pos, self.pgresult.ntuples, self._make_row)
    self._pos = self.pgresult.ntuples
    return records
```

We have the assert but sometimes the last line would fail because
`self.pgresult` was None, which can only happen when the `Cursor` is
used concurrently.

This solution is somewhat questionable because it replaces the original
design where we have `count` tasks with a design where we have
`concurrency` tasks but each task does `count / concurrency` operations.
An alternative would be to implement a thread safe pool of Cursors, say
using a `threading.Lock` and a `list`, but that is more work and
complexity.

I did check the results of the benchmark before and after this change
and they look roughly the same.