expr: guard against reducing `to_jsonb(bool): jsonb` to `_: boolean`
In certain parts of the `MirScalarExpr` handling code we trigger term
simplification when a term is known to be a `boolean` literal.

Unfortunately, due to the lack of a first-class internal representation
for `Jsonb` data, functions like `MirScalarExpr::is_literal_true` will
return `Datum::True` for both of the following cases:

```sql
SELECT true           -- input is literal `true: boolean`
SELECT to_jsonb(true) -- input is literal `true: jsonb`
```

Consequently, we might simplify a (well-typed) `jsonb` term into a term
of type `boolean`. In order to avoid that, certain places where this
might happen must be guarded with additional check of the term type.

Fixes #21244.