adapter: don't double terminate on startup fail
handle_startup was incorrectly handling the case where it successfully
completed a connection startup, but the AdapterClient had gone away
before the Coordinator could respond to it. This could happen if the
pgwire, http, or websocket connection was closed during handle_startup's
execution. When handle_startup observed that failure, it called
handle_terminate. But when the AdapterClient was dropped, it also
terminated the connection. This double termination caused envd to panic
because that should never happen.

Fix this by not even creating an AdapterClient until after startup is
complete. That way there's no Drop implementation until we have a known
good session. This should also fix a previous, unsolved bug with
handle_terminate that was likely happening for the same reason.

Fixes #21409