storage: teach load generator sources to resume
The current load generator framework requires each implementation to
always produce its differential history from the beginning of time. The
framework relied on this behavior to be able to count up the number of
transactions already emitted and provide a definite behavior. This is
suboptimal however as many load generator implementations would rather
skip forward to the resumption point without generating all this data
for it to be immediately discarded.

This PR teaches the load generator framework about resumption in the
form of a `resume_offset` argument passed to anyone implementing the
`Generator` trait. The expectation is that the source is obligated to
produce all the updates that happen beyond `resume_offset` and may skip
producing the updates that happen not beyond `resume_offset`. The latter
is not an obligations and for sources that is beneficial for
impementation reasons to reproduce their history they are free to do so.