explain: fix `export_ids_for` helper function
It turns out that transient IDs are only used for dataflows resulting
from `CREATE MATERIALIZED VIEW` statements and never occur in
dataflows resulting from `CREATE INDEX`.

Change the `export_ids_for` to account for that and add some extensive
comments documenting why this implementation is in line with the current
dataflow construction behavior.