explain: move `broken` modifier to `Explainee::Query`
The `BROKEN` modifier is only meaningful if the explainee is a `Query`,
because this is the only explainee variant for which we run the query
pipeline (as opposed to looking up a preiously saved plan).

It therefore makes more sense to move the representation of this
modifier in our inernal structures from the enclosing `ExplainPlan~`
structure to `Explainee`.