Encode some datums in a variable-length way (#21016)
Slightly belated part 2 of
https://github.com/MaterializeInc/materialize/pull/4646

This adds some new tag variants for timestamps as well as signed ints,
allowing us to encode them in less space in most cases. For timestamps,
we only use the full 16-byte representation if we can't fit in 8 bytes.
For integers, we write down only the amount of bytes we actually need.

There are two things we could tweak to get even more savings:

1. We could do this for unsigned integers as well. I didn't bother
because AFAIK they are rare enough that it wasn't worth consuming a
bunch more tag space.
2. Currently we still use one bit to encode the sign, which roughly
halves the set of values representable at each level. We could encode it
in the tag instead (doubling the amount of new tags).