Add memory savings to mz_expected_group_size_advice
This commit adds a new column to the introspection view
mz_expected_group_size_advice providing an estimate of the memory
savings that can be expected by applying the tuning hint. The memory
savings are a conservative estimate based on the sizes of the input
arrangements for each level to be cut. These arrangements should
dominate the size of each level that can be cut, since the reduction
gadget internal to the level does not remove much data at these levels.