rbac: Fix privilege requirement for create source
The cluster configuration of a CreateSourcePlan dictates whether the
source will be created in an existing cluster, create a new cluster, or
do the default behavior, which is usually to create a new cluster.
However, if the source is a progress source or a sub-source, then the
default behavior is to do whatever the main source does.

RBAC mistakenly interpreted the cluster configurations of progress
and sub-sources to mean that a new cluster was always created and
required the CREATECLUSTER system privilege.

This commit fixes the described confusion so the CREATECLUSTER system
privilege is only required when a cluster is created.

Fixes #21543