storage: generator: only emit messages beyond resume frontier
The work to add resumption to generators implemented in #21482 had the
unforeseen sideeffect that all the data emitted by generators would
enter the dataflow. Since some generators always generate all of their
data from the beginning of time this could lead to a large amount of
data being emitted without yielding, since yielding only happens after
we've reached the resumption frontier.

To address this issue this commit filters the messages that are not
beyond the resumption frontier.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>