storage_state: remove assertion on suspend commands
SuspendAndRestart commands are sent concurrently with controller
commands and their interleaving doesn't have any useful property to
assert. In particular it is entirely possible to receive a
`SuspendAndRestart` command for a source that has already been dropped.
The current code is set up to panic in this case even though this is not
necessarily evidence of a bug.

This commit removes the assertion.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>