compute-client: cancel served peeks
This commit makes the compute client explictly send `CancelPeek`
commands for peeks it doesn't need a response for anymore (because it
already saw one). This allows late replicas to clean up any state they
hold for this peek. It will also allow us to simplify peek handling in
the future.