compute*: simplify peek reduction in the history
Now that we send a `CancelPeek` command for every finished peek, we can
simplify how peek reduction works in the `ComputeCommandHistory`.
Previously, callers had to pass a list of peeks that should still be
considered active to `push`, to ensure `Peek` commands would get cleaned
up during reduction. Instead we can now remove `Peek`s when we see
corresponding `CancelPeek` commands. This results in a cleaner API for
the `ComputeCommandHistory`.