compute-client: only wait for first peek response
This commit simplifies the peek tracking in the compute controller by
changing the logic to only wait for the first applicable peek response.

Upon receiving that first peek response, the compute controller
immediately stops tracking the peek and drops the read hold it has
installed for it. This is safe because the compute protocol allows
sending `Peek` commands for times that have already been compacted. Such
peeks may produce errors or invalid rows, but that doesn't matter
because we ignore their responses anyway.

This commit also moves the emission of `CancelPeek` commands into
`remove_peek`. This ensures that whenever the controller stops tracking
a peek, that peek is also explictily canceled and callers don't have to
be careful to ensure this.