compute-client: clean up replica-targeted peeks
When a replica is dropped, we need to check if there are any pending
peeks that target this replica and remove those, so we don't leave them
pending forever. We return a `PeekResponse::Error` for pending peeks
dropped that way, to make the caller aware.

Note that this logic will also abort peeks when the targeted replica
fails rather than being dropped. This commit takes the stance that this
behavior is fine because (a) the caller is made aware of the unhealthy
replica this way and (b) this behavior is consistent with how we handle
targeted subscribes. In contrast to subscribes it would be safe to not
abort targeted peeks on replica failure, but the implementation would be
more complicated because we would need to start differenciating between
permanent and temporary removal of replicas.