storage: clean up source metadata handling
This PR disentangles a longstanding "abstraction leak" in the ingestion
pipeline where Kafka specific details had spread to the entirety of the
pipeline. Specifically, the structs involved in the ingestion pipeline
(`SourceMessage`, `SourceOutput`, `DecodeResult`) all carried kafka
specific information (partition, offset, headers, upstream_time) which
forced all other source types to provide dummy values for them. Also,
because the metadata production (inside the source) from metadata
consumption (in the decode operator) was separate we had a lot more code
to ensure that metadata is only produced when it is to be consumed
downstream.

All this is fixed by moving the Kafka speicific details where they
belong, i.e inside the Kafka source. The various structs now carry a
single `metadata: Row` field which is produced by the source only when
there is metadata to produce, and its schema can be determined by
looking at the corresponding source connetion struct.

The only Kafka specific field left in there is the position offset,
which unfortunately is currently needed in the upsert operator. A future
PR will clean this up too.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>