Revised docs for new group size hints
This commit revises the user-facing documentation to refer to the new
group size hints, namely AGGREGATE INPUT GROUP SIZE, DISTINCT ON INPUT
GROUP SIZE, and LIMIT INPUT GROUP SIZE, instead of the old syntax with
the EXPECTED GROUP SIZE. Even though the old syntax is still available
for backwards compatibility, the documentation is being changed to only
refer to the new syntax. As we help customers move to the new syntax,
we will eventually be in a position where the old syntax can be
deprecated. This strategy was proposed in design doc #21500.