catalog: Remove SerializedX items from catalog (#21562)
Previously, the catalog had structs of the form 'SerializedX', such as
SerializedItem. They were used to persist catalog objects to the stash
while allowing us to evolve the on disk state independently of the in
memory state.
    
Now that we use protobuf to persist objects to the stash, the
SerializedX structs are no longer necessary. This commit removes these
structs from the catalog. Some of the structs are still necessary to
use as arguments for catalog::storage functions, so those were moved
to the catalog::storage module.