catalog: Refactor storage timeline dependencies
This commit refactors the mz_adapter::catalog::storage dependencies so
that it no longer relies on mz_adapter::timeline. This will help in the
effort to extract catalog::storage into its own crate.

The logic for reserving a new timestamp range is moved from
mz_adapter::catalog::storage::Connection::open to
mz_adapter::catalog::Catalog::open. As a result, the timestamp used for
catalog migrations may be ahead of the timestamp used for the stash
initialization. This is OK because the two steps are logically
different and do not happen atomically. It was already possible for
this to happen due to a crash in-between the two steps.

Works towards resolving #20953