rbac: Consolidate RBAC requirement generation
This commit consolidates the code responsible for generating the
requirements needed to execute a plan. Previously there was a separate
function and match statement for each type of requirement. Those are:

  - role membership
  - object ownership
  - object privileges
  - superuser
  - required item usage

This had the following drawbacks:

  - The total set of requirements needed to execute a plan were spread
  across multiple functions, making it difficult to see the entire
  picture for a single plan.
  - Updating/creating a new plan required updating many different
  match statements.
  - Anecdotally, if there are multiple match statements each with a
  large default branch, people are encouraged to add new plans
  to the end of each default branch without thinking about it.

This commit combines all plan match statements into a single function
and match statement that generates all requirements for a single plan.
This is purely code movement and hasn't changed any RBAC related logic.