adapter: Clean up peeks on dropped views
During a DDL statement the adapter will check and see if there's a
pending peek or subscribe that depends on the dropped object. If any
dependent peeks or subscribe exist, then they will be cancelled.
Previously, the logic for this did not check for dropped views. This
commit updates the logic so that views are included.

Fixes #21497