rbac: Consolidate RBAC requirement generation (#21618)
This commit consolidates the code responsible for generating the
requirements needed to execute a plan. Previously there was a separate
function and match statement for each type of requirement. Those are:

  - role membership
  - object ownership
  - object privileges
  - superuser
  - required item usage

This had the following drawbacks:

- The total set of requirements needed to execute a plan were spread
across multiple functions, making it difficult to see the entire picture
for a single plan.
- Updating/creating a new plan required updating many different match
statements.
- Anecdotally, if there are multiple match statements each with a large
default branch, people are encouraged to add new plans to the end of
each default branch without thinking about it.

This commit combines all plan match statements into a single function
and match statement that generates all requirements for a single plan.
This is purely code movement and hasn't changed any RBAC related logic.

I'm also curious to hear if people think this is a better structure or
the old structure was better.

### Motivation
This PR refactors existing code.

### Tips for Reviewer

The content of this PR is mostly code movement and re-structuring. I
would recommend doing the following:
- Look at the new struct `RbacRequirements`.
- Look at the changes in the function `check_plan`.
- Check the signatures and first couple of match branches of the
functions that were removed:
  - `generate_required_role_membership`
  - `generate_required_ownership`
  - `generate_required_privileges`
  - `check_superuser_required`
- Look at the signature and first couple of match branches in
`generate_rbac_requirement`

The intention is that `generate_rbac_requirement` is a combination of
all the removed methods. So you shouldn't have to closely analyze the
contents of the removed functions or the contents of
`generate_rbac_requirement`, since it's meant to be purely code
movement. Of course if you'd like you can do a careful comparison.