transform: struct-based `DerivedAttributes` container
The `DerivedAttributes` and `RequiredAttributes` structs were hosting
instances of the attribute derivation algorithms for the attributes to
be derived in `TypeMap` instances.

This is preventing us from seeding the algorithm (`Attribute`) instances
with `&dyn ExprHumanizer` context because `TypeMap` keys need to be
thread-safe (`Send + Sync`).

In order to allow `Attribute` implementations to depend on context types
such as `&dyn ExprHumanizer` or `&dyn StatisticsOracle` we replace the
`TypeMap` fields with a custom struct called `AttributeStore`.

The struct has an `Option<T>` field provisioned for each `A` that
implements `Attribute`. Fields are listed in dependency order and
methods that need to respect the dependency order access them in the
same sequence. An implementation of `AttributeContainer<A>` for each
field provides a unified interfacing for accessin the attribute of type
`A` from client code. See the rustdoc comments on these items for more
details.

The change removes the following code from the `attribute` framework:

- The `typemap_rev` crate dependency (which provided `TypeMap`).
- The `AsKey` helper struct.
- The `AttributeId` enum and the `TOTAL_ATTRIBUTES` const field.