Make statement logging persistent (#21379)
This stores the statement log in storage-managed collections, rather
than in system tables, so they won't be refreshed and cleared on
startup. In order to avoid unnecessarily high load on Cockroach, we
buffer the updates and flush them every 5 seconds.

On startup, we re-ingest the log in order to truncate the parts of it
that are older than a configurable value, by default 30 days. Currently,
this has the potential to use significant memory; we may need to switch
to a more incremental approach. We will need to watch this closely when
turning on statement logging in prod.

### Motivation

  * This PR adds a known-desirable feature.

    persisted statement logging

### Tips for Reviewer

By far the most complicated part of this is
`partially_truncate_statement_logging`. Most of the rest should be
straightforward.

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly
considered.
- [x] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [x] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [x] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - Save values in in the statement log for 30 days.