storage/kafka: snapshot topic on initialization
This PR is changing the Kafka source implementation such that it
produces a snapshot of the whole topic that will get reclocked to a
single Materialize timestamp. This ensures that if we ingest a topic on
which key compaction has taken place users will not be able to observe
incorrect results which can arrise when viewing only part of the
compacted prefix.

TODO: Explain better, add example

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>