rbac: relax RBAC requirements for `EXPLAIN <query>`
At the moment, executing `EXPLAIN <query>` statements requires read
privileges to all items referenced in the query.

Since `EXPLAIN` is not evaluating a query (except for partially
evaluating constant paths), we can relax the RBAC requirements to only
require `USAGE` privilege on the schema of items contained in the query.

Closes #20478.