explain: don't swallow regular tracing calls
The current `OptimizerTrace` implementation substitutes the default
tracing `Subscriber` with a brand new one containing only `PlanTrace`
layers. This means that regular `tracing::event!` calls are not
propagated to the regularly configured subscribers, which can be a
problem if one tries to rely on them for debugging code that lives
inside `explain_query_optimizer_pipeline`.

This PR fixes that limitation by introducing a `DelegateSubscriber` that
can be used to delegate subscriber calls to an inner subscriber. The new
subscriber type is used as the base of the `OptimizerTrace` subscriber
chain instead of `Registry::default()`.