adapter: Apply write ts in bootstrapping
Previously, the Coordinator would do the following during bootstrap:

  1. Get a write timestamp from the timestamp oracle, ws1.
  2. Advance all tables to ws1.
  3. Get a read timestamp from the timestamp oracle, rs1.
  4. Read a snapshot of all system tables at rs1.
  5. Get a write timestamp from the timestamp oracle, ws2.
  6. Write retractions for all system tables at ws2.
  7. Apply ws2 with the timestamp oracle.

This commit adds a step after step (2) but before step (3) to apply ws1
with the timestamp oracle. This guarantees that rs1 >= ws1, which was
not previously true. There shouldn't be any existing issues from this
since the table advancements are all empty writes, but it's still a
better practice to always apply write timestamps after using them.