storage: start tables with an initial since
Present tables as springing into existence at the register_ts by
advancing the since. Otherwise, we could end up in a situation where a
table with a long compaction window appears to exist before the
environment (and this the table) existed.

We could potentially also do the same thing for other sources, in
particular storage's internal sources and perhaps others, but leave them
for now.