Add types to AvailableCollections
This commit captures the MIR-level typing information and saves it in
the AvailableCollections struct of LIR-level Plan::ArrangeBy nodes.
Note that capturing type information in this fashion is quadratic in
the number of nodes in the MIR expression, which is unfortunate.
However, the commit illustrates that capturing the type information is
possible for further use. A possible solution to improve the
performance issue is memoization.